Health replaces the score: matches are won by draining the rival
Both sides now have health in every match. A goal takes the ball's
points times the rules' GoalDamage, and ApplyDamage is open for spells.
The seven-segment numbers show health (rounded up) and flash on damage.
Rules hold StartingHealth (0 = endless, counting goals). In a run the
player fights on the run's health, which every hit comes off.

+++ Source/IJPong/Run/IJPActConfig.h
@@ -122,7 +122,7 @@ public:
 
 	/** Health restored at a Rest node (never above the run's maximum). */
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "Act|Nodes", meta = (ClampMin = "0"))
-	int32 RestHeal = 3;
+	float RestHeal = 3.f;
 
 	/** The encounter for a fight node type, or null for Rest (and the unbuilt Shop / Event). */
 	const FIJPEncounter* GetEncounter(EIJPNodeType Type) const
